bau: Remove tests until needed
Currently this project is in an exploration phase. As such,
comprehensive unit and e2e tests are not strictly necessary
for prototyping work.

They are also not currently part of our build pipeline, so
are not providing us the protection that would be normally
expected from them.

These tests are being removed for the time being, until a
time that they can be properly used and developed. At that
point git history can be used, and required files can be
easily restored.